# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pamir Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: getProfiles
      tags: [profiles]
      summary: Get all profiles
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileInput"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get all posts (feed)
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: groupId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /posts/{id}/likes:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeInput"
      responses:
        "200":
          description: Like result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeInput"
      responses:
        "200":
          description: Unlike result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{id}/comments:
    get:
      operationId: getComments
      tags: [posts]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [posts]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        institution:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        badge:
          type: string
          nullable: true
        rating:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - rating
        - createdAt
        - updatedAt
    CreateProfileInput:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        institution:
          type: string
        city:
          type: string
        bio:
          type: string
        badge:
          type: string
      required:
        - name
    UpdateProfileInput:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
        institution:
          type: string
        city:
          type: string
        bio:
          type: string
        badge:
          type: string
        rating:
          type: integer
    Post:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        groupId:
          type: string
          nullable: true
        likes:
          type: integer
        liked:
          type: boolean
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
      required:
        - id
        - author
        - content
        - likes
        - liked
        - comments
        - createdAt
    CreatePostInput:
      type: object
      properties:
        authorId:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        groupId:
          type: string
      required:
        - authorId
        - content
    Comment:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        text:
          type: string
        createdAt:
          type: string
      required:
        - id
        - author
        - text
        - createdAt
    CreateCommentInput:
      type: object
      properties:
        authorId:
          type: string
        text:
          type: string
      required:
        - authorId
        - text
    LikeInput:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    LikeResult:
      type: object
      properties:
        likes:
          type: integer
        liked:
          type: boolean
      required:
        - likes
        - liked
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        fromUser:
          $ref: "#/components/schemas/UserProfile"
        text:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - fromUser
        - text
        - read
        - createdAt